refactor(search): move query mapping into the module
Move buildSearchFilters/buildSearchSort and the SearchQueryFilters shape into
the search module (utils.ts/types.ts) so the /store/search route stays thin —
it just reads req.filterableFields + pagination and delegates the mapping.
